Step: Derived Trait Calculation

Calculate glycan derived traits using glydet::derive_traits(). Advanced glycan structure analysis that summarizes structural properties of a glycome or each glycosite. Need glycan structure information.

Usage

step_derive_traits(trait_fns = NULL, mp_fns = NULL, mp_cols = NULL)

Arguments

trait_fns

A named list of derived trait functions created by trait factories. Names of the list are the names of the derived traits. Default is NULL, which means all derived traits in basic_traits() are calculated.

mp_fns

A named list of meta-property functions. This parameter is useful if your trait functions use custom meta-properties other than those in all_mp_fns(). Default is NULL, which means all meta-properties in all_mp_fns() are used.

mp_cols

A character vector of column names in the var_info tibble to use as meta-properties. If names are provided, they will be used as names of the meta-properties, otherwise the column names will be used. Meta-properties specified in mp_cols will overwrite those introduced by mp_fns with the same names, including the built-in meta-properties. Default is NULL, which means no columns are used as meta-properties.

Value

A glysmith_step object.

Details

Data required:

exp: The experiment to calculate derived traits for

Data generated:

trait_exp: The experiment with derived traits

Tables generated:

derived_traits: A table containing the derived traits.

See also

Examples

step_derive_traits()
#> <step "step_derive_traits()"> Derived trait calculation