Step: Preprocessing

Preprocess the experiment using glyclean::auto_clean(). This step can be omitted if the experiment is already preprocessed.

Usage

step_preprocess(
  batch_col = "batch",
  qc_name = "QC",
  normalize_to_try = NULL,
  impute_to_try = NULL,
  remove_preset = "discovery",
  batch_prop_threshold = 0.3,
  check_batch_confounding = TRUE,
  batch_confounding_threshold = 0.4
)

Arguments

batch_col

The column name in sample_info for batches. Default is "batch". Can be NULL when no batch information is available.

qc_name

The name of QC samples in the group_col column. Default is "QC". Only used when group_col is not NULL.

normalize_to_try

Normalization functions to try. A list. Default includes:

normalize_median(): median normalization

normalize_median_abs(): absolute median normalization

normalize_total_area(): total area mormalization

normalize_quantile(): quantile normalization

normalize_loessf(): LoessF normalization

normalize_loesscyc(): LoessCyc normalization

normalize_rlr(): RLR normalization

normalize_rlrma(): RLRMA normalization

normalize_rlrmacyc(): RLRMAcyc normalization

impute_to_try

Imputation functions to try. A list. Default includes:

impute_zero(): zero imputation

impute_sample_min(): sample-wise minimum imputation

impute_half_sample_min(): half sample-wise minimum imputation

impute_sw_knn(): sample-wise KNN imputation

impute_fw_knn(): feature-wise KNN imputation

impute_bpca(): BPCA imputation

impute_ppca(): PPCA imputation

impute_svd(): SVD imputation

impute_min_prob(): minimum probability imputation

impute_miss_forest(): MissForest imputation

remove_preset

The preset for removing variables. Default is "discovery". Available presets:

"simple": remove variables with more than 50% missing values.

"discovery": more lenient, remove variables with more than 80% missing values, but ensure less than 50% of missing values in at least one group.

"biomarker": more strict, remove variables with more than 40% missing values, and ensure less than 60% of missing values in all groups.

batch_prop_threshold

The proportion of variables that must show significant batch effects to perform batch correction. Default is 0.3 (30%).

check_batch_confounding

Whether to check for confounding between batch and group variables. Default to TRUE.

batch_confounding_threshold

The threshold for Cramer's V to consider batch and group variables highly confounded. Only used when check_batch_confounding is TRUE. Default to 0.4.

Value

A glysmith_step object.

Details

Data required:

exp: The experiment to preprocess

Data generated:

raw_exp: The raw experiment (previous exp, saved for reference)

This step is special in that it silently overwrites the exp data with the preprocessed experiment. This ensures that no matter if preprocessing is performed or not, the "active" experiment is always under the key exp. The previous exp is saved as raw_exp for reference.

See also

Examples

step_preprocess()
#> <step "step_preprocess()"> Preprocessing
step_preprocess(remove_preset = "discovery")
#> <step "step_preprocess(remove_preset = \"discovery\")"> Preprocessing